refactor: split chrome ui
