+++ src/egui_app/ui/chrome/mod.rs
@@ -1,0 +1,4 @@
+mod buttons;
+mod status_area;
+mod top_bar;
+mod tool_panels;
